+++ src/components/SearchMap/ReusableMapContainer.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { node, string } from 'prop-types';
-import mapValues from 'lodash/mapValues';
+import { node, string, object } from 'prop-types';
 import { IntlProvider } from 'react-intl';
-import messages from '../../translations/en.json';
 import config from '../../config';
 
 import css from './SearchMap.css';
@@ -60,15 +58,8 @@ class ReusableMapContainer extends React.Component {
     // NOTICE: Children rendered with ReactDOM.render doesn't have Router access
     // You need to provide onClick functions and URLs as props.
     const renderChildren = () => {
-      const isTestEnv = process.env.NODE_ENV === 'test';
-
-      // Locale should not affect the tests. We ensure this by providing
-      // messages with the key as the value of each message.
-      const testMessages = mapValues(messages, (val, key) => key);
-      const localeMessages = isTestEnv ? testMessages : messages;
-
       const children = (
-        <IntlProvider locale={config.locale} messages={localeMessages}>
+        <IntlProvider locale={config.locale} messages={this.props.messages}>
           {this.props.children}
         </IntlProvider>
       );
@@ -128,6 +119,7 @@ ReusableMapContainer.propTypes = {
   children: node.isRequired,
   className: string,
   reusableMapHiddenHandle: string.isRequired,
+  messages: object.isRequired,
 };
 
 export default ReusableMapContainer;