+++ src/util/test-helpers.js
@@ -8,6 +8,10 @@ import en from 'react-intl/locale-data/en';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import configureStore from '../store';
+
+// In case you have translated the template and have new translations that
+// are missing from the en translations file, the language for the tests can
+// be changed here so that there are no missing translation keys in tests.
 import messages from '../translations/en.json';
 
 Enzyme.configure({ adapter: new Adapter() });